refactor(invoices): Introduce lib/commons
Provides interfaces, fn resolvePayment, fn getCountryCode, fn getReference and fn getHrId.

These were previosuly in lib/paymentslip and might be used beyond paymetslip purposes.
