+++ packages/invoices/index.ts
@@ -1,1 +1,2 @@
+export * as commons from './lib/commons'
 export * as paymentslip from './lib/paymentslip'
