+++ packages/republik-crowdfundings/lib/payments/paymentslip/sendPaymentReminders.ts
@@ -8,7 +8,7 @@ import {
 import { formatPrice } from '@orbiting/backend-modules-formats'
 import { Context } from '@orbiting/backend-modules-types'
 import { publishFinance } from '@orbiting/backend-modules-republik/lib/slack'
-import { paymentslip } from '@orbiting/backend-modules-invoices'
+import * as invoices from '@orbiting/backend-modules-invoices'
 
 const { sendMailTemplate } = require('@orbiting/backend-modules-mail')
 const { PAYMENT_DEADLINE_DAYS } = require('./helpers')
@@ -229,7 +229,7 @@ async function send({
     `api/email/payment/reminder${isLast ? '/last' : ''}/default/subject`,
   ])
 
-  const resolvedPayment = await paymentslip.resolve(
+  const resolvedPayment = await invoices.commons.resolvePayment(
     { hrid: payment.hrid },
     context,
   )
@@ -253,18 +253,23 @@ async function send({
         },
         {
           name: 'reference',
-          content: paymentslip.getReference(resolvedPayment.hrid, true),
+          content: invoices.commons.getReference(resolvedPayment.hrid, true),
         },
       ],
       attachments: [
         {
           type: 'application/pdf',
-          name: [
-            'QR-Rechnung',
-            creditor?.address?.name,
-            paymentslip.getReference(resolvedPayment.hrid, true)
-          ].filter(Boolean).join(' ') + '.pdf',
-          content: (await paymentslip.generate(resolvedPayment)).toString('base64'),
+          name:
+            [
+              'QR-Rechnung',
+              creditor?.address?.name,
+              invoices.commons.getReference(resolvedPayment.hrid, true),
+            ]
+              .filter(Boolean)
+              .join(' ') + '.pdf',
+          content: (
+            await invoices.paymentslip.generate(resolvedPayment)
+          ).toString('base64'),
         },
       ],
     },
