+++ packages/republik-crowdfundings/lib/scheduler/payments/parseCamt053.ts
@@ -1,7 +1,7 @@
 import parser from 'fast-xml-parser'
 import { Nominal } from 'simplytyped'
 
-import { paymentslip } from '@orbiting/backend-modules-invoices'
+import * as invoices from '@orbiting/backend-modules-invoices'
 
 type Amount = Nominal<string, 'amount'>
 type AmountInCents = Nominal<number, 'cents'>
@@ -45,7 +45,7 @@ type TransactionDetails = {
       CdtrRefInf?: {
         Ref?: string
       }
-    } 
+    }
   }
 }
 
@@ -250,7 +250,8 @@ function getMitteilung(
   transactionDetails?: TransactionDetails,
 ): string | null {
   const creditorReference = transactionDetails?.RmtInf?.Strd?.CdtrRefInf?.Ref
-  const referenceHrId = creditorReference && paymentslip.getHrId(creditorReference)
+  const referenceHrId =
+    creditorReference && invoices.commons.getHrId(creditorReference)
 
   if (referenceHrId) return referenceHrId
 
@@ -264,9 +265,11 @@ function getMitteilung(
   }
 
   return (
-    paymentslip.getHrId(remittanceInformation) ||
+    invoices.commons.getHrId(remittanceInformation) ||
     remittanceInformation?.match(/\b([A-Za-z0-9]{6})\b/)?.[1] ||
-    avisierungstext.match(/.*?MITTEILUNGEN:.*?\s([A-Za-z0-9]{6})(\s.*?|$)/)?.[1] ||
+    avisierungstext.match(
+      /.*?MITTEILUNGEN:.*?\s([A-Za-z0-9]{6})(\s.*?|$)/,
+    )?.[1] ||
     null
   )
 }
